Try to install appimagelauncher without scripts.
Explanation: The appimagelauncher.rpm file ships a script that will run
(postinstall phase) a command to ensure appiamges are opened using
`appimagelauncher`. It uses `update-binfmts`, but this fails because it
needs a kernel module (-> uses modprobe) to do it. We can safely disable
this script, because the file also ships a rule in
`/usr/lib/binfmt.d/appimage.conf`, and systemd should be able to use
this file to add the rule at boot time.